+++ DfE.FindInformationAcademiesTrusts.Data/Repositories/Ofsted/IOfstedRepository.cs
@@ -3,4 +3,6 @@ namespace DfE.FindInformationAcademiesTrusts.Data.Repositories.Ofsted;
 public interface IOfstedRepository
 {
     Task<AcademyOfsted[]> GetAcademiesInTrustOfstedAsync(string uid);
+    Task<OfstedInspectionHistorySummary> GetOfstedInspectionHistorySummaryAsync(int urn);
+    Task<OfstedShortInspection> GetOfstedShortInspectionAsync(int urn);
 }