+++ DfE.FindInformationAcademiesTrusts/Pages/Schools/SchoolNavMenu.cs
@@ -2,9 +2,12 @@ using DfE.FindInformationAcademiesTrusts.Configuration;
 using DfE.FindInformationAcademiesTrusts.Data.Enums;
 using DfE.FindInformationAcademiesTrusts.Extensions;
 using DfE.FindInformationAcademiesTrusts.Pages.Schools.Contacts;
+using DfE.FindInformationAcademiesTrusts.Pages.Schools.Governance;
+using DfE.FindInformationAcademiesTrusts.Pages.Schools.Ofsted;
 using DfE.FindInformationAcademiesTrusts.Pages.Schools.Overview;
 using DfE.FindInformationAcademiesTrusts.Pages.Shared.NavMenu;
 using Microsoft.FeatureManagement;
+using GovernanceAreaModel = DfE.FindInformationAcademiesTrusts.Pages.Schools.Governance.GovernanceAreaModel;
 
 namespace DfE.FindInformationAcademiesTrusts.Pages.Schools;
 
@@ -26,7 +29,11 @@ public class SchoolNavMenu(IVariantFeatureManager featureManager) : ISchoolNavMe
         [
             GetServiceNavLinkTo<OverviewAreaModel>(OverviewAreaModel.PageName, "/Schools/Overview/Details",
                 activePage),
-            GetServiceNavLinkTo<ContactsAreaModel>(ContactsAreaModel.PageName, contactLink, activePage)
+            GetServiceNavLinkTo<ContactsAreaModel>(ContactsAreaModel.PageName, contactLink, activePage),
+            GetServiceNavLinkTo<GovernanceAreaModel>(GovernanceAreaModel.PageName, "/Schools/Governance/Current",
+                activePage),
+            GetServiceNavLinkTo<OfstedAreaModel>(OfstedAreaModel.PageName, "/Schools/Ofsted/SingleHeadlineGrades",
+                activePage)
         ];
     }
 
@@ -46,6 +53,8 @@ public class SchoolNavMenu(IVariantFeatureManager featureManager) : ISchoolNavMe
         {
             OverviewAreaModel => BuildLinksForOverviewPage(activePage),
             ContactsAreaModel => await BuildLinksForContactsPageAsync(activePage),
+            GovernanceAreaModel governanceAreaModel => BuildLinksForGovernancePage(governanceAreaModel),
+            OfstedAreaModel => BuildLinksForOfstedPage(activePage),
             _ => throw new ArgumentOutOfRangeException(nameof(activePage), activePage, "Page type is not supported.")
         };
     }
@@ -107,6 +116,30 @@ public class SchoolNavMenu(IVariantFeatureManager featureManager) : ISchoolNavMe
                 inSchoolNavLink
             ]
             : [inSchoolNavLink];
+    }
+
+    private static NavLink[] BuildLinksForGovernancePage(GovernanceAreaModel activePage)
+    {
+        var currentNavLink = GetSubNavLinkTo<CurrentModel>(GovernanceAreaModel.PageName,
+            CurrentModel.NavTitle(activePage.SchoolGovernance.Current.Length),
+            "/Schools/Governance/Current", activePage,
+            "current-governors-subnav");
+
+        var historicNavLink = GetSubNavLinkTo<HistoricModel>(GovernanceAreaModel.PageName,
+            HistoricModel.NavTitle(activePage.SchoolGovernance.Historic.Length),
+            "/Schools/Governance/Historic", activePage,
+            "historic-governors-subnav");
+
+        return [currentNavLink, historicNavLink];
+    }
+
+    private static NavLink[] BuildLinksForOfstedPage(ISchoolAreaModel activePage)
+    {
+        return
+        [
+            GetSubNavLinkTo<SingleHeadlineGradesModel>(OfstedAreaModel.PageName, SingleHeadlineGradesModel.SubPageName,
+                "/Schools/Ofsted/SingleHeadlineGrades", activePage, "ofsted-single-headline-grades-subnav")
+        ];
     }
 
     private static NavLink GetSubNavLinkTo<T>(string serviceName, string linkDisplayText, string aspPage,