+++ tests/DFE.FindInformationAcademiesTrusts.CypressTests/cypress/pages/trusts/trustContactsPage.ts
@@ -163,7 +163,7 @@ class ContactsPage {
         const { datasource } = this.elements;
         datasource.section().expandDetailsElement();
         datasource.section().should('have.attr', 'open');
-        datasource.trmLatestUpdatedBy().should('contain.text', 'Trust relationship manager was updated by ' + expectedUser + ' on ');
+        datasource.trmLatestUpdatedBy().should('include.text', 'Trust relationship manager updated by ' + expectedUser + ' on ');
         return this;
     }
 
@@ -171,7 +171,7 @@ class ContactsPage {
         const { datasource } = this.elements;
         datasource.section().expandDetailsElement();
         datasource.section().should('have.attr', 'open');
-        datasource.sfsoLatestUpdatedBy().should('contain.text', 'SFSO (Schools financial support and oversight) lead was updated by ' + expectedUser + ' on ');
+        datasource.sfsoLatestUpdatedBy().should('include.text', 'SFSO (Schools financial support and oversight) lead updated by ' + expectedUser + ' on ');
         return this;
     }