+++ CHANGELOG.md
@@ -9,6 +9,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Added Manage projects and cases
+- 
+## [Release-44][release-44] (production-2025-08-06.6211)
+
+### Changed
+
+- Added Ofsted single headline grades page for a school or academy
+
+## [Release-43][release-43] (production-2025-07-31.6191)
+
+### Changed
+
+- Added current and historic governors for schools and academies
 
 ## [Release-42][release-42] (production-2025-07-17.6155)
 
@@ -414,7 +426,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Updated text on "404 - Not Found" page to adhere to DfE design pattern and added the not found url to the support email template.
 
 [unreleased]:
-  https://github.com/DFE-Digital/find-information-about-academies-and-trusts/compare/production-2025-07-17.6155...HEAD
+  https://github.com/DFE-Digital/find-information-about-academies-and-trusts/compare/production-2025-08-06.6211...HEAD
 [release-1]:
   https://github.com/DFE-Digital/find-information-about-academies-and-trusts/releases/tag/production-2024-07-18.2517
 [release-2]:
@@ -499,3 +511,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   https://github.com/DFE-Digital/find-information-about-academies-and-trusts/releases/tag/production-2025-07-09.6117
 [release-42]:
   https://github.com/DFE-Digital/find-information-about-academies-and-trusts/releases/tag/production-2025-07-17.6155
+[release-43]:
+  https://github.com/DFE-Digital/find-information-about-academies-and-trusts/releases/tag/production-2025-07-31.6191
+[release-44]:
+  https://github.com/DFE-Digital/find-information-about-academies-and-trusts/releases/tag/production-2025-08-06.6211